Update battle and summary screen configurations for type icons and shadow styles

+++ include/graphics.h
@@ -1980,7 +1980,7 @@ extern const u32 gSummaryPage_Skills_Tilemap[];
 extern const u32 gSummaryPage_BattleMoves_Tilemap[];
 extern const u32 gSummaryPage_ContestMoves_Tilemap[];
 extern const u16 gPPTextPalette[];
-extern const u16 gMoveTypes_Pal[];
+extern const u32 gMoveTypes_Pal[];
 extern const u32 gMoveTypes_Pal_New[];
 
 extern const u16 gSummaryScreen_MoveEffect_Battle_Tilemap[];
